requirements.txt
boto3==1.34.75
face_recognition==1.3.0
flet==0.21.2
flet_core==0.21.2
flet_material==0.3.3
flet_route==0.3.2
numpy==1.26.4
opencv_python==4.9.0.80
Pillow==10.3.0
python-dotenv==1.0.1
Requests==2.31.0
ujson==5.9.0
